Add AnnotateRoutes::BaseProcessor#strip_annotations and #real_content_and_header_position / Remove them from AnnotateRoutes::Helpers
